+++ both/layouts/Navigation.jsx
@@ -34,7 +34,7 @@ FacilityFilter = React.createClass({
             {
                 facility?
                 <div style={{position:"absolute",top:"-14px",left:"-8px",whiteSpace:"nowrap"}}>
-                    <img style={{width:"40px",float:"left",display:"inline-block","borderRadius":"2px"}} alt="image" src={facility.thumb[0]} />
+                    <img style={{width:"40px",float:"left",display:"inline-block","borderRadius":"2px"}} alt="image" src={facility.getThumbUrl()} />
                     <div style={{"whiteSpace":"normal",width:"165px","display":"inline-block","fontSize":"12px","lineHeight":"13px","padding":"8px 0 0 5px"}} className="nav-label">{facility.getName()}</div>
                 </div>
                 :