+++ both/views/Facility/FacilityEdit.jsx
@@ -259,15 +259,13 @@ FacilityEdit = React.createClass({
 		if(facility){
 			team = facility.getTeam();
 			contacts = facility.getContacts();
-			if(team) {
-				tenants = team.getMembers();
-				return {
-					ready:true,
-					facility:facility,
-					contacts:contacts,
-					schema:schema,
-					tenants:tenants.slice(0,3),
-				}
+			tenants = facility.getTenants();
+			return {
+				ready:true,
+				facility:facility,
+				contacts:contacts,
+				schema:schema,
+				tenants:tenants
 			}
 		}
 		return {
@@ -282,6 +280,11 @@ FacilityEdit = React.createClass({
 		return function() {
 			item.save();
 		}
+	},
+
+	updateField(field,value) {
+		this.props.item[field] = value;
+		this.props.item.save();
 	},
 
 	form1 : ["name","address"],
@@ -314,10 +317,16 @@ FacilityEdit = React.createClass({
 				        <AutoForm item={facility} schema={schema} form={['_attachments']} save={this.save()} />
 					</CollapseBox>
 			   		<CollapseBox title="Contacts">
-			   			<ContactList items={contacts} />
+			   			<ContactList 
+			   				items={contacts} 
+			   				onChange={this.updateField.bind(null,'_contacts')}
+			   			/>
 					</CollapseBox>
 			   		<CollapseBox title="Tenants">
-			   			<ContactList items={tenants} />
+			   			<ContactList 
+			   				items={tenants} 
+			   				onChange={this.updateField.bind(null,'_tenants')}
+			   			/>
 					</CollapseBox>
 			   		<CollapseBox title="Facility holder" collapsed={true}>
 				        <div className="col-lg-12" style={{paddingTop:"20px"}}>